+++ src/python/pants/backend/go/goals/tailor_test.py
@@ -4,7 +4,7 @@ import textwrap
 
 import pytest
 
-from pants.backend.go import module, target_type_rules
+from pants.backend.go import target_type_rules
 from pants.backend.go.goals.tailor import (
     PutativeGoExternalModuleTargetsRequest,
     PutativeGoModuleTargetsRequest,
@@ -12,7 +12,7 @@ from pants.backend.go.goals.tailor import (
 )
 from pants.backend.go.goals.tailor import rules as go_tailor_rules
 from pants.backend.go.target_types import GoExternalPackageTarget, GoModule, GoPackage
-from pants.backend.go.util_rules import external_module, sdk
+from pants.backend.go.util_rules import external_module, go_mod, sdk
 from pants.core.goals.tailor import (
     AllOwnedSources,
     PutativeTarget,
@@ -33,7 +33,7 @@ def rule_runner() -> RuleRunner:
             *go_tailor_rules(),
             *external_tool.rules(),
             *source_files.rules(),
-            *module.rules(),
+            *go_mod.rules(),
             *external_module.rules(),
             *sdk.rules(),
             *target_type_rules.rules(),
