+++ src/python/pants/backend/go/pkg_integration_test.py
@@ -4,7 +4,7 @@ import textwrap
 
 import pytest
 
-from pants.backend.go import module, pkg
+from pants.backend.go import pkg
 from pants.backend.go.pkg import (
     ResolvedGoPackage,
     ResolveExternalGoModuleToPackagesRequest,
@@ -12,7 +12,7 @@ from pants.backend.go.pkg import (
     ResolveGoPackageRequest,
 )
 from pants.backend.go.target_types import GoModule, GoPackage
-from pants.backend.go.util_rules import external_module, sdk
+from pants.backend.go.util_rules import external_module, go_mod, sdk
 from pants.build_graph.address import Address
 from pants.core.util_rules import external_tool, source_files
 from pants.engine.fs import EMPTY_DIGEST
@@ -26,7 +26,7 @@ def rule_runner() -> RuleRunner:
         rules=[
             *external_tool.rules(),
             *source_files.rules(),
-            *module.rules(),
+            *go_mod.rules(),
             *external_module.rules(),
             *pkg.rules(),
             *sdk.rules(),
