+++ src/python/pants/backend/go/target_type_rules_test.py
@@ -4,7 +4,7 @@ import textwrap
 
 import pytest
 
-from pants.backend.go import module, pkg, target_type_rules
+from pants.backend.go import pkg, target_type_rules
 from pants.backend.go.target_type_rules import InferGoPackageDependenciesRequest
 from pants.backend.go.target_types import (
     GoExternalPackageTarget,
@@ -13,7 +13,7 @@ from pants.backend.go.target_types import (
     GoPackage,
     GoPackageSources,
 )
-from pants.backend.go.util_rules import external_module, sdk
+from pants.backend.go.util_rules import external_module, go_mod, sdk
 from pants.build_graph.address import Address
 from pants.core.util_rules import external_tool, source_files
 from pants.engine.addresses import Addresses
@@ -35,7 +35,7 @@ def rule_runner() -> RuleRunner:
         rules=[
             *external_tool.rules(),
             *source_files.rules(),
-            *module.rules(),
+            *go_mod.rules(),
             *pkg.rules(),
             *external_module.rules(),
             *sdk.rules(),
